% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot.R
\name{lytics_theme}
\alias{lytics_theme}
\title{Lytics ggplot2 theme}
\usage{
lytics_theme()
}
\description{
\code{lytics_theme} is a ggplot2 theme based on internal Lytics styleguides.
}
\details{
The theme can be added to any ggplot plot
}
\examples{
qplot(...) + lytics_theme()
}

