% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot.R
\name{set_lytics_theme}
\alias{set_lytics_theme}
\title{Load the Lytics ggplot2 theme}
\usage{
set_lytics_theme()
}
\description{
\code{lytics_theme} is a ggplot2 theme based on internal Lytics styleguides.
}
\details{
Unfortunately, you can't set aesthetic defaults in ggplot2 themes.
}
\examples{
set_lytics_theme()
}

